Extract Precedence to its own module
